pyproject.toml
[tool.poetry]
name = "opensr_test"
version = "1.3.3"
description = "A comprehensive benchmark for real-world Sentinel-2 imagery super-resolution"
authors = ["Cesar Aybar <cesar.aybar@uv.es>"]
repository = "https://github.com/csaybar/opensr-test"
documentation = "https://csaybar.github.io/opensr-test/"
readme = "README.md"
packages = [
  {include = "opensr_test"}
]

[tool.poetry.dependencies]
python = ">=3.10, <4.0"
satalign = ">=0.1.3"
numpy = ">=1.25.2"
matplotlib = ">=3.7.0"
torchvision = ">=0.17.0"
scikit-image = ">=0.19.0"
requests = ">=2.25.0"
kornia = ">=0.7.2"
pydantic = ">=2.7.1"
opencv-python = ">=4.8.0.0"
lpips = {extras = ["perceptual"], version = ">=0.1.4"}
openai-clip = {extras = ["perceptual"], version = ">=1.0.1"}
open-clip-torch = {extras = ["perceptual"], version = ">=2.24.0"}
mpltern = ">=1.0.4"



[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
pytest-cov = "^4.0.0"
deptry = "^0.12.0"
mypy = "^1.5.1"
pre-commit = "^3.4.0"
tox = "^4.11.1"

[tool.poetry.group.docs.dependencies]
mkdocs = "^1.4.2"
mkdocs-material = "^9.2.7"
mkdocstrings = {extras = ["python"], version = "^0.23.0"}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
files = ["opensr_test"]
disallow_untyped_defs = "True"
disallow_any_unimported = "True"
no_implicit_optional = "True"
check_untyped_defs = "True"
warn_return_any = "True"
warn_unused_ignores = "True"
show_error_codes = "True"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py37"
line-length = 120
fix = true
select = [
    # flake8-2020
    "YTT",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-comprehensions
    "C4",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # mccabe
    "C90",
    # pycodestyle
    "E", "W",
    # pyflakes
    "F",
    # pygrep-hooks
    "PGH",
    # pyupgrade
    "UP",
    # ruff
    "RUF",
    # tryceratops
    "TRY",
]
ignore = [
    # LineTooLong
    "E501",
    # DoNotAssignLambda
    "E731",
]

[tool.ruff.format]
preview = true

[tool.coverage.report]
skip_empty = true

[tool.coverage.run]
branch = true
source = ["opensr_test"]


[tool.ruff.per-file-ignores]
"tests/*" = ["S101"]
